break out FileUploadArea
